+++ invoice-backend/src/services/user.service.ts
@@ -1,25 +1,27 @@
 import { inject, injectable } from "inversify";
-import { PrismaUserRepo } from "../repositories/implementations/prismaUserRepo";
 import {
   UserEntity,
   CreateUserDTO,
   UserDTO,
-  LoginResponseDTO,
+  UserIdAndRole,
 } from "../constants/types";
 import bcrypt from "bcrypt";
 import { IUserRepo } from "../repositories/userRepo";
 import { validateUserCreate, validateUserList } from "../utils/utils";
-import { SECRET } from "../config/server.config";
-import jwt from "jsonwebtoken";
 import {
   InternalServerException,
   NotFoundException,
-  UnauthorizedException,
+  ValidationException,
 } from "../config/exception.config";
+import TYPES from "@/constants/identifiers";
 
 @injectable()
 export class UserService {
-  constructor(@inject(PrismaUserRepo) private readonly userRepo: IUserRepo) {}
+  constructor(
+    @inject(TYPES.IUserRepo) private readonly userRepo: IUserRepo,
+    @inject(TYPES.UserContext)
+    private readonly userContext: UserIdAndRole,
+  ) {}
 
   createUser = async (args: CreateUserDTO): Promise<UserDTO> => {
     const validatedArgs = validateUserCreate(args);
@@ -36,15 +38,37 @@ export class UserService {
     const userDTO: UserDTO = {
       id: createdUser.id,
       name: createdUser.name,
-      username: createdUser.username,
+      username: createdUser.username ?? "",
+    };
+
+    return userDTO;
+  };
+
+  createUserWithAuth0 = async (args: UserIdAndRole): Promise<UserIdAndRole> => {
+    const createdUser = await this.userRepo.createUserWithAuth0(args);
+
+    const userDTO: UserIdAndRole = {
+      id: createdUser.id,
+      name: createdUser.name,
+      username: createdUser.username ?? "",
+      role: createdUser.role,
     };
 
     return userDTO;
   };
 
   getUsers = async () => {
+    if (!this.userContext) {
+      throw new ValidationException("Unauthorized");
+    }
+
+    const { role, id } = this.userContext;
+
+    if (!id) {
+      throw new ValidationException("Unauthorized");
+    }
     try {
-      const userList = await this.userRepo.findAllUsers();
+      const userList = await this.userRepo.getAllUsers();
       const validatedUserList: UserDTO[] = validateUserList(userList);
       return validatedUserList;
     } catch (error) {
@@ -53,8 +77,17 @@ export class UserService {
     }
   };
 
-  getUser = async (id: number) => {
-    const user = await this.userRepo.findUserById(id);
+  getUser = async (id: string) => {
+    if (!this.userContext) {
+      throw new ValidationException("Unauthorized");
+    }
+
+    const { role } = this.userContext;
+
+    if (!id) {
+      throw new ValidationException("Unauthorized");
+    }
+    const user = await this.userRepo.getUserById(id);
 
     if (!user) {
       throw new NotFoundException("User not found");
@@ -68,39 +101,38 @@ export class UserService {
     return userDTO;
   };
 
-  login = async (
-    username: string,
-    password: string,
-  ): Promise<LoginResponseDTO> => {
-    const user = await this.userRepo.findUserByUsername(username);
-
-    if (!user) {
-      throw new NotFoundException("User not found");
-    }
-
-    const match = await bcrypt.compare(password, user.passwordHash);
-
-    if (!match) {
-      throw new UnauthorizedException("Invalid username or password");
-    }
-
-    const tokenPayload = {
-      id: user.id,
-      username: user.username,
-    };
-
-    const userDTO: UserDTO = {
-      id: user.id,
-      name: user.name,
-      username: user.username,
-    };
-
-    const token = jwt.sign(tokenPayload, SECRET, { expiresIn: "1h" });
-    const loginResponse: LoginResponseDTO = { token: token, user: userDTO };
-    return loginResponse;
+  getUserByIdSafely = async (id: string) => {
+    const user = await this.userRepo.getUserByIdSafely(id);
+    return user;
   };
 
   deleteUsers = async (): Promise<boolean> => {
+    if (!this.userContext) {
+      throw new ValidationException("Unauthorized");
+    }
+
+    const { role, id } = this.userContext;
+
+    if (!id) {
+      throw new ValidationException("Unauthorized");
+    }
     return await this.userRepo.deleteAllUsers();
+  };
+
+  deleteUsersKeepAdmin = async (): Promise<boolean> => {
+    if (!this.userContext) {
+      throw new ValidationException("Unauthorized");
+    }
+
+    const { role, id } = this.userContext;
+
+    if (!id) {
+      throw new ValidationException("Unauthorized");
+    }
+
+    if (role !== "ADMIN") {
+      throw new ValidationException("Unauthorized - must be ADMIN");
+    }
+    return await this.userRepo.deleteAllUsersKeepAdmin();
   };
 }
